requirements.txt
fastapi>=0.111
uvicorn>=0.30
langchain>=0.2
langchain-openai>=0.1
langchain-community>=0.2
langgraph>=0.2
langgraph-checkpoint-postgres>=3.0
psycopg[binary]>=3.3
psycopg-pool>=3.3
python-dotenv>=1.0
